(PUP-8950) Move handle_connection_errors to Puppet::Util::SSL
