+++ lib/puppet/util/ssl.rb
@@ -1,6 +1,8 @@
+require 'openssl'
+
 ##
 # SSL is a private module with class methods that help work with x.509
-# subjects.
+# subjects and errors.
 #
 # @api private
 module Puppet::Util::SSL
@@ -49,5 +51,42 @@ module Puppet::Util::SSL
 
   def self.is_possibly_valid_dn?(dn)
     dn =~ /=/
+  end
+
+  ##
+  # Extract and format meaningful error messages from OpenSSS::OpenSSLErrors
+  # and a Validator. Re-raises the error if unknown.
+  #
+  # @api private
+  #
+  # @param [OpenSSL::OpenSSLError] error An error thrown during creating a
+  #   connection
+  # @param [Puppet::SSL::DefaultValidator] verifier A Validator who may have
+  #   invalidated the connection
+  # @param [String] host The DNS name of the other end of the SSL connection
+  #
+  # @raises [Puppet::Error, OpenSSL::OpenSSLError]
+  def self.handle_connection_error(error, verifier, host)
+    # can be nil
+    peer_cert = verifier.peer_certs.last
+
+    if error.message.include? "certificate verify failed"
+      msg = error.message
+      msg << ": [" + verifier.verify_errors.join('; ') + "]"
+      raise Puppet::Error, msg, error.backtrace
+    elsif peer_cert && !OpenSSL::SSL.verify_certificate_identity(peer_cert, host)
+      valid_certnames = [peer_cert.subject.to_s.sub(/.*=/, ''),
+                         *Puppet::SSL::Certificate.subject_alt_names_for(peer_cert)].uniq
+      if valid_certnames.size > 1
+        expected_certnames = _("expected one of %{certnames}") % { certnames: valid_certnames.join(', ') }
+      else
+        expected_certnames = _("expected %{certname}") % { certname: valid_certnames.first }
+      end
+
+      msg = _("Server hostname '%{host}' did not match server certificate; %{expected_certnames}") % { host: host, expected_certnames: expected_certnames }
+      raise Puppet::Error, msg, error.backtrace
+    else
+      raise error
+    end
   end
 end
